+++ controllers/saisieCom/commandeController.js
@@ -2,6 +2,8 @@
 const Individu = require('../../models/individu');
 const Article = require('../../models/article');
 const Commande = require('../../models/commande');
+const Anomalie = require('../../models/anomalie')
+const {calculPrix, verifSoldeCB, testAnomalie } = require('./functionAnomalies')
 
 const commande_getOne = async (req, res) => {
     try {
@@ -13,6 +15,43 @@ const commande_getOne = async (req, res) => {
     } catch (err) {
         console.log(err);
     };
+}
+
+const commande_verif = async (req, res) => {
+    let commande = await Commande.findById(req.params.id)
+    //pour récupérer la liste des prix des articles de notre commande
+    const ids = commande.articles;
+    const articles = await Article.find({ _id: { $in: ids } });
+    const lprix = [];
+    ids.forEach(id => {
+        articles.forEach(article => {
+            if (article.id == id) {
+                lprix.push(article.prix);
+            }
+        });
+    });
+
+    commande.prix = calculPrix(lprix, commande.quantite);
+    if(commande.pCarte=='on' && commande.numeroCarte!="" && commande.dateExpiration!="" && commande.crypto!=""){
+        commande.valeur=verifSoldeCB(commande);
+    }
+    commande.etat = testAnomalie(commande);
+
+    if(commande.etat.length>0){
+        const anomalie=new Anomalie();
+        anomalie.numeroCom=commande.numCommande;
+        anomalie.client=commande.client;
+        anomalie.anomalies=commande.etat;
+        anomalie.save();
+    }
+    commande.verification = true;
+    commande.save()
+        .then((result) => {
+            res.redirect('/creerCom');
+        })
+        .catch((err) => {
+            console.log(err);
+        });
 }
 
 const commande_delete = (req, res) => {
@@ -28,5 +67,6 @@ const commande_delete = (req, res) => {
 
 module.exports = {
     commande_getOne,
+    commande_verif,
     commande_delete
 }
