+++ views/partials/barreSaisieCom.ejs
@@ -1,4 +1,4 @@
 <div class="barre">
     <input class="Nouvel_Commande" id="New_Comm" type="submit" value="Créer une nouvelle Commande" onclick="location.href='/creerCom';">
-    <input class="Recherche_Commande" id="R_Comm" type="submit" value="Compléter une Commande" onclick="location.href='/modifCom';">
+    <input class="Recherche_Commande" id="R_Comm" type="submit" value="Vérifier validité d'une commande" onclick="location.href='/modifCom';">
 </div>
